+++ examples/byob/variables.tf
@@ -9,12 +9,12 @@ variable "location" {
 }
 
 variable "prefix" {
+  description = "Storage Account prefix"
   type        = string
-  description = "Prefix of the storage account and storage container."
 }
 
 variable "deletion_protection" {
-  description = "If the instance should have deletion protection enabled. The storage container can't be deleted when this value is set to `true`."
+  description = "Storage Account delete protection"
   type        = bool
-  default     = true
+  default     = false
 }