+++ modules/byob/main.tf
@@ -1,8 +1,10 @@
 module "storage" {
   source = "../storage"
 
-  create_queue = false
-  namespace    = var.prefix
+  create_queue        = false
+  namespace           = var.prefix
+  resource_group_name = var.resource_group_name
+  location            = var.location
 
   deletion_protection = var.deletion_protection
 }