+++ CHANGELOG.md
@@ -2,6 +2,40 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.3.0](https://github.com/wandb/terraform-azurerm-wandb/compare/v1.2.1...v1.3.0) (2023-04-03)
+
+
+### Features
+
+* Enable non azure bucket ([#15](https://github.com/wandb/terraform-azurerm-wandb/issues/15)) ([2939eed](https://github.com/wandb/terraform-azurerm-wandb/commit/2939eedf4a101116c7301f6bad0af3883a07e7c4))
+
+
+### Bug Fixes
+
+* Enable non azure bucket ([7f1b9ba](https://github.com/wandb/terraform-azurerm-wandb/commit/7f1b9ba558d783d3db65f313f1f729834ca502f5))
+
+### [1.2.1](https://github.com/wandb/terraform-azurerm-wandb/compare/v1.2.0...v1.2.1) (2022-11-07)
+
+
+### Bug Fixes
+
+* Fix byob logic for queue ([94a27ad](https://github.com/wandb/terraform-azurerm-wandb/commit/94a27add94e1f7b0f1b06a04750849cdc26ca4fe))
+
+## [1.2.0](https://github.com/wandb/terraform-azurerm-wandb/compare/v1.1.0...v1.2.0) (2022-11-02)
+
+
+### Features
+
+* BYOB ([375d83b](https://github.com/wandb/terraform-azurerm-wandb/commit/375d83bc9a43e051844b02864be346d8e298106a))
+* BYOB ([9e4fecc](https://github.com/wandb/terraform-azurerm-wandb/commit/9e4fecc240ce8f50992490bf5550bb8adff6d586))
+
+## [1.1.0](https://github.com/wandb/terraform-azurerm-wandb/compare/v1.0.1...v1.1.0) (2022-10-21)
+
+
+### Features
+
+* Add storage account creds ([2d543be](https://github.com/wandb/terraform-azurerm-wandb/commit/2d543bea5562e28591da6e49fa57e98d83d2fa92))
+
 ### [1.0.1](https://github.com/wandb/terraform-azurerm-wandb/compare/v1.0.0...v1.0.1) (2022-09-02)